refactor(app): extract preferences sanitizers from usePreferences god file
Move ~260 lines of pure sanitization/migration logic out of
usePreferences.ts into app/stores/preferences/sanitizers.ts:
- legacy migration detection
- plain-object/blocked-browser-object guards
- selection string and optional enum normalizers
- snapshot cloning + persisted state parser

Continues the decompose-god-files theme; usePreferences.ts shrinks by
~220 lines and now imports the sanitizers as a focused module.
